Fix vessel recovery handling for experience.
When vessel recovery is requested (ie, the event fires), the game does not
get saved before the next load (it gets saved beforehand), so the change in
experience tracking does not get saved. However, some time after the
processing event fires, the game is saved, so do the handling in there.