Delay actual loading of the scenario data.
As part of the process of loading, locations are checked and for that, the
vessels need to have been loaded, but it appears vessel loading comes after
scenario loading, so delay a frame.

Also, to help make Progeny more usable, use an event to signal that the
scenario has been loaded.